<frame> and <frameset> are not allowed in Outlook Edition of Word 2007, remove them
The following tags are reported as valid by the Outlook 2007 HTML and CSS Validator tool but are not actually permitted in Outlook 2007 e-mail because of security concerns:

Frameset

Frame

https://learn.microsoft.com/en-us/previous-versions/office/developer/office-2007/aa338201(v=office.12)#outlook-2007-security-concerns

These tags are valid for Word 2007 content, but are intentionally blocked when it is used by Outlook 2007 because of security implications. Microsoft reserves the right to block certain tags that might compromise the privacy of users or the security of the e-mail content in which the content is displayed.